+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,16 @@
 
 Release History
 ===============
+
+0.2.7
++++++
+* remove client side sku check for linux app service plan create
+
+0.2.6
++++++
+* update ACR SDK
+* webapp: fix a bug in `az webapp config backup update` that prevents setting a backup schedule if one is not already set
+
 0.2.5
 +++++
 * az functionapp create supports creating a linux consumption plan type with a specific runtime