+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+2.2.6
++++++
+* Minor changes
+
+2.2.5
++++++
+* Minor changes
 
 2.2.4
 +++++